refactor : Ladderfactory interface 삭제 및 이름 변경
