+++ src/main/java/nextstep/ladder/LadderMain.java
@@ -3,7 +3,6 @@ package nextstep.ladder;
 import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.domain.Players;
 import nextstep.ladder.factory.LadderFactory;
-import nextstep.ladder.factory.RandomLadderFactory;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.OutputView;
 
@@ -12,8 +11,8 @@ public class LadderMain {
         Players players = new Players(InputView.inputPerson());
         int height = InputView.inputLadderHeight();
 
-        LadderFactory ladderFactory = new RandomLadderFactory();
-        Ladder ladder = ladderFactory.createLadder(height, players.countPerson());
+        LadderFactory ladderFactory = new LadderFactory();
+        Ladder ladder = ladderFactory.createRandomLadder(height, players.countPerson());
 
         OutputView.outputResult(players, ladder);
     }
